Video Recording/Main: More optimizations for recording older guests which have a slightly different screen update logic [build fix]. bugref:10650
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@105096 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-client/RecordingStream.cpp
@@ -643,7 +643,7 @@ int RecordingStream::SendVideoFrame(PRECORDINGVIDEOFRAME pVideoFrame, uint64_t m
         offDst += cbDstBytesPerLine;
         offSrc += pVideoFrame->Info.uBytesPerLine;
     }
-    pFrame->u.Video.Info.uBytesPerLine = cbDstBytesPerLine;
+    pFrame->u.Video.Info.uBytesPerLine = (uint32_t)cbDstBytesPerLine;
 
     pFrame->enmType     = RECORDINGFRAME_TYPE_VIDEO;
     pFrame->msTimestamp = msTimestamp;
